+++ src/components/groups/GroupListItem/GroupListItem.component.tsx
