+++ src/components/groups/GroupListItem/index.ts
@@ -1,0 +1,1 @@
+export * from './GroupListItem.component';
